Create vlans_per_device.py
NetMRI Client script to print all the VLANs on the device.  In this case it prints the VLAN number, then VLAN name.  I assume the VLAN information in NetMRI is the same regardless of the vendor, however, this was tested with Juniper devices.  Juniper devices all have VLAN names, so if there is no VLAN name in NetMRI, you may get a "record-not-found" response back.  Just remove the val.VlanName from the print statement.